Revert "Mi 49/registration page"
