+++ package.json
@@ -39,7 +39,6 @@
         "vue-jest": "^3.0.5",
         "vue-loader": "^15.7.1",
         "vue-style-loader": "^4.1.2",
-        "vue-svg-loader": "^0.12.0",
         "vue-template-compiler": "^2.6.10",
         "webpack": "^4.41.0",
         "webpack-cli": "^3.3.9",
