+++ src/frontend/App.vue
@@ -1,27 +1,32 @@
 <template>
-    <main>        
+    <main>
+        <nav-bar />
         <router-view />
     </main>
 </template>
 
-<script>    
+<script>
+    import NavBar from './components/NavBar.vue';
+
+    export default {
+        components: {
+            'nav-bar': NavBar
+        }
+    }
 </script>
 
 <style lang="scss">
-    @import url(https://fonts.googleapis.com/css?family=Arimo|DM+Serif+Display|Poppins:700&display=swap);
-
     * {
         font-family: sans-serif;
         box-sizing: border-box;
     }
 
-    html, body {
+    body {
         padding: 0;
         margin: 0;
-        height: 100%;
     }
 
-    main{
-        height: 100%;
+    article {
+        text-align: center;
     }
 </style>
