+++ src/frontend/index.js
@@ -1,34 +1,34 @@
 
-import Vue from 'vue';
-import VueRouter from 'vue-router';
+import Vue from 'vue'
+import VueRouter from 'vue-router'
 import App from './App';
 
-import RegistrationPage from './pages/RegistrationPage';
+import WelcomePage from './pages/WelcomePage';
 import HelpPage from './pages/HelpPage';
 import NotFoundPage from './pages/NotFoundPage';
 
-Vue.use(VueRouter);
+Vue.use(VueRouter)
 
 const router = new VueRouter({
     mode: 'history',
     routes: [
-        {
-            path: '/',
-            component: RegistrationPage
+        { 
+            path: '/', 
+            component: WelcomePage 
+        },
+        { 
+            path: '/help', 
+            component: HelpPage 
         },
         {
-            path: '/help',
-            component: HelpPage
-        },
-        {
-            path: '*',
+            path :'*',
             component: NotFoundPage
         }
     ]
-});
+})
 
 new Vue({
     el: '#app',
     render: h => h(App),
     router
-});
+})
